+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Added additional GROQ examples
 * Added a slug uniqueness validator for document level translations (`import { isSlugUnique } from '@sanity/document-internationalization/lib/validators`)
 * Fixed issue with `fixIdStructureMismatchDocuments` maintenance function
+* `delimiter` is now the default ID structure
+* Added maintenance function to add base document ref to translated documents
 
 <!-- // @TODO remove old changelog? -->
 ## v5.6.0